+++ app/(client-app)/user/dashboard/page.tsx
@@ -3,5 +3,10 @@
 import type { JSX } from 'react';
 
 export default function DashboardPage(): JSX.Element {
-  return <div>Dashboard</div>;
+  return (
+    <div>
+      <h1>Dashboard</h1>
+      <p>Paragraph about dashboard.</p>
+    </div>
+  );
 }