+++ app/(client-app)/user/layout.tsx
@@ -2,10 +2,18 @@
 
 import type { JSX, ReactNode } from 'react';
 
+import BottomNavWrapper from '@/components/bottomNav/BottomNavWrapper';
 import RequiresSignedInGuard from '@/guards/RequiresSignedInGuard';
 
 export default function UserLayout({
   children,
 }: Readonly<{ children: ReactNode }>): JSX.Element {
-  return <RequiresSignedInGuard>{children}</RequiresSignedInGuard>;
+  return (
+    <RequiresSignedInGuard>
+      <div className="h-full flex flex-col">
+        <div className="flex-grow min-h-0">{children}</div>
+        <BottomNavWrapper />
+      </div>
+    </RequiresSignedInGuard>
+  );
 }